Fix PushTopic subscribers (#4)
* correct the upsert! arguments

* Use a new Restforce client when starting event machine.

While manually testing the EM loop, the subscribers were not receiving
any pushtopic events.  A simple script that started an EM loop with a
single subscriber to the same Sf instance and topic name worked as
expected.

There was a difference in the output of `lsof` between the
two running programs. The script had 2 TCP file descriptors open and
connected to a port on an SF host while the server only had 1.  After
swapping out the reused SalesforceStreamer for a fresh client, the
pushtopic events began flowing as expected. Not sure the root cause but
likely has something to do with the the Restforce client not being
thread safe. Afterall, the maintainers do mention in the README of
restforce/restforce:

> It is also important to note that the client object should not be
> reused across different threads, otherwise you may encounter
> thread-safety issues.

* use correct config syntax in readme

* no use for checking diff of name since its the lookup key

* fix push event INFO logger