Use SETEX Redis command
The previous implementation of the RedisReplay class used the sorted set
operations to gaurantee that the highest messageId would be sent to
Salesforce Streaming API upon booting up.

If that messageId is older than 24 hours, then no messages will be received
from the Salesforce Streaming API.  This is problematic when a
Salesforce PushTopic doesn't receive events very frequently and can lead
to the server doing nothing without warning/errors to the operator.

It's more important that we ensure new events are received by the push
topic than it is to ensure we don't replay a message we already
received.  This is a big assumption that might not be true for all
applications, but for all applications not receiving new events is bad.

The new implementation uses the SETEX command to set the messageId for
the PushTopic with an expiration of 24 hours. This way when the server
reboots after no events for > 24 hours, then the GET command will
retrieve nothing and a -1 will be used to start fetching new events.

The side effect is that if messages are received out of order before the
server restarts, then one or more messages already received may be
replayed.