Do not memoize the response from ReplayPersistence.retrieve (#20)
* Do not memoize the response from ReplayPersistence.retrieve

The streamer process records the replayId received from each event. The
algorithm to resubscribe to a PushTopic uses the value returned by
PushTopic#replay. This value was being memoized, and because the
PushTopic object is only initialized once when the streamer process is
started up the PushTopic#replay value becomes stale.

This change incurs some overhead because each subscribe event will not
read from the ReplayPersistence (default Redis) but it will now return
the most recently received replayId.

Without this change, the stale replayId could be as stale as 24 hours or
more. If a subscription to a PushTopic uses a replayId that is older
than the message durability retention period specified by the Salesforce
Streaming API docs then no messages (past, present, or future!) are sent
to the subscribing client.  The current restforce client will receive a
400 error message from the Salesforce API, but do nothing and the
connection and streamer process will appear alive and well but unable to
receive new events!

* bump patch version to release urgent fix