Adds ReplayIdErrorExtension to raise exception
The Restforce client uses the Faye library to communicate via CometD
with the Salesforce Streaming API. The Restforce Faye client does not
handle application errors from the Streaming API. Particularly if a bad
replayId is given.

This commit adds an extension to the Restforce Faye client to look for
these application specific errors. If a "400" replayId error is found in
the message, then an exception is raised so that the process exits
non-zero.