adds middleware pattern to push topic handler
This commit introduces the middleware pattern so that we can configure a
series of handlers in the client to support more custom use cases that
don't fit into the this gem.

Configure the middleware stack globally when booting up your server using

    Configuration.instance.use_middleware MyMiddleware

Then when the push topic receives a message the chain will be used to
pass the message through each.